[OVZC-55] 매핑 수정
