+++ src/main/java/kr/ac/kgu/kpserver/domain/board/Articles/ArticleService.java
@@ -4,6 +4,7 @@ import kr.ac.kgu.kpserver.domain.board.Articles.dto.CreateArticleRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.webjars.NotFoundException;
 
 @Service
 @RequiredArgsConstructor
@@ -23,6 +24,13 @@ public class ArticleService {
     @Transactional
     public void deleteArticle(Article article) {
         articleRepository.delete(article);
+    }
+
+    @Transactional
+    public int getLikesForArticle(Long articleId) throws NotFoundException {
+        Article article = articleRepository.findById(articleId)
+                .orElseThrow(() -> new NotFoundException("Could not find article with id : " + articleId));
+        return article.getLikes().size();
     }
 
 }
